fix ColorPicker not displaying correct integer values; fix room json ui elements

+++ mod-sdk/src/main/java/com/github/argon/sos/mod/sdk/data/IntegerValue.java
@@ -42,7 +42,7 @@ public class IntegerValue {
         return getValue() <= min;
     }
 
-    public double inc(int value) {
+    public int inc(int value) {
         return setValue(getValue() + value);
     }
 }
